try fix

+++ research/slim/train_image_classifier.py
@@ -449,7 +449,7 @@ def main(_):
       logits, end_points = network_fn(images)
 
       logits = tf.Print(logits, [logits, end_points['Predictions']], 'lg,pr', summarize=60)
-      softmax = tf.losses.softmax_cross_entropy(labels, logits, loss_collection=None)
+      softmax = tf.losses.softmax_cross_entropy(tf.cast(labels, tf.float32), logits, loss_collection=None)
       logits = tf.Print(logits, [softmax], 'softmax', summarize=60)
 
       #############################
